// DoAnXepLich/frontend/src/types/hocKy.ts
export interface HocKy {
  maHocKy: number
  maNamHoc: number
  tenNamHoc?: string
  tenHocKy: string
  ngayBatDau: string
  ngayKetThuc: string
  trangThai: string
}

export interface TaoHocKyPayload {
  maNamHoc: number
  tenHocKy: string
  ngayBatDau: string
  ngayKetThuc: string
  trangThai: string
}